Finally I finished likebutton system

+++ html2/com_hate.php
@@ -5,9 +5,20 @@ if (isset($_SESSION['nickname'])) {
     $checkifalreadyquery = mysqli_query($conn, $checkifalready);
     $checkifalreadyrow = mysqli_fetch_array($checkifalreadyquery);
     if (isset($checkifalreadyrow)) {
-        echo "<script>alert('이미 좋아요 또는 싫어요를 누른 게시물입니다.')</script>";
-        echo "<meta http-equiv='Refresh' content=\"0; url='com_show.php?id={$_SESSION['id']}'\" />";
-    } else {
+        if ($checkifalreadyrow[3] == 'hate') {
+            $lessgood = "UPDATE com SET hate=hate-1 where id={$_SESSION['id']}";
+            mysqli_query($conn, $lessgood);
+            $removefromlikebutton = "DELETE from likebutton where postid = {$_SESSION['id']} and nickname='{$_SESSION['nickname']}'";
+            mysqli_query($conn, $removefromlikebutton);
+            echo "<meta http-equiv='Refresh' content=\"0; url='com_show.php?id={$_SESSION['id']}'\" />";
+        } elseif ($checkifalreadyrow[3] == 'like') {
+            $changelikeandhate = "UPDATE com SET hate=hate+1 , good=good-1 where id={$_SESSION['id']}";
+            mysqli_query($conn, $changelikeandhate);
+            $changefromlikebutton = "UPDATE likebutton SET type='hate' where postid={$_SESSION['id']} and nickname='{$_SESSION['nickname']}'";
+            mysqli_query($conn, $changefromlikebutton);
+            echo "<meta http-equiv='Refresh' content=\"0; url='com_show.php?id={$_SESSION['id']}'\" />";
+        };
+    } elseif (!isset($checkifalreadyrow)) {
         $morehate = "UPDATE com SET hate=hate+1 where id={$_SESSION['id']}";
         mysqli_query($conn, $morehate);
         $sql = "SELECT * from com where id={$_SESSION['id']}";
@@ -16,7 +27,7 @@ if (isset($_SESSION['nickname'])) {
         $addtolikelist = "INSERT into likebutton (nickname, postid, type, title, written) VALUES ('{$_SESSION['nickname']}', {$_SESSION['id']}, 'hate', '{$row[1]}', '{$row[3]}')";
         mysqli_query($conn, $addtolikelist);
         echo "<meta http-equiv='Refresh' content=\"0; url='com_show.php?id={$_SESSION['id']}'\" />";
-    }
+    };
 } else {
     echo "<script>alert('로그인 후 이용 가능합니다')</script>";
     echo "<meta http-equiv='Refresh' content=\"0; url='com_show.php?id={$_SESSION['id']}'\" />";
